Added LuckPerms for Admin commands

+++ BungeeFeatures/src/main/java/de/joshiworld/bungee/commands/deletewarp.java
@@ -1,5 +1,6 @@
 package de.joshiworld.bungee.commands;
 
+import de.joshiworld.api.LuckPermsAPI;
 import de.joshiworld.bungee.Bungee;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.CommandSender;
@@ -11,15 +12,17 @@ public class deletewarp extends Command {
     public deletewarp() {
         super("deletewarp");
     }
+    private final LuckPermsAPI luckPerms= new LuckPermsAPI(Bungee.getInstance());
+
     @Override
     public void execute(CommandSender sender, String[] args) {
         if(!(sender instanceof ProxiedPlayer)) return;
         ProxiedPlayer player = (ProxiedPlayer) sender;
+
+
         String uuid = player.getUniqueId().toString();
-        if(args.length==0) {
-            player.sendMessage( new TextComponent(ChatColor.RED.toString()+"Specify a Warp"));
-            return;
-        }
+        if(args.length == 0 || !(luckPerms.hasPermissionGroup("bungeefeatures.warp",player.getUniqueId())))return;
+
         if(Bungee.getInstance().data.searchHome("warp",args[0])){
             Bungee.getInstance().data.deleteHome("warp",args[0]);
             player.sendMessage(new TextComponent(ChatColor.RED.toString()+"Warp deleted Successfully"));
